updat

+++ src/main/java/com/isoft/controller/WzController.java
@@ -3,9 +3,12 @@ package com.isoft.controller;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.isoft.dao.InfosMapper;
+import com.isoft.dao.UserMapper;
 import com.isoft.dao.WzMapper;
 import com.isoft.pojo.entity.Infos;
+import com.isoft.pojo.entity.User;
 import com.isoft.pojo.entity.Wz;
+import com.isoft.pojo.vo.WzVo;
 import com.isoft.service.InfosService;
 import com.isoft.service.WzService;
 import com.isoft.utils.ResponseData;
@@ -19,51 +22,72 @@ import java.util.Map;
 
 /**
  * <p>
- *  前端控制器
+ * 前端控制器
  * </p>
- *
  */
-@Controller
+@RestController
 @RequestMapping("/wz")
 public class WzController {
 
-	@Autowired
-	private WzService wzService;
+    @Autowired
+    private WzService wzService;
 
-	@Autowired
-	private WzMapper wzMapper;
+    @Autowired
+    private WzMapper wzMapper;
 
-	// 添加
-	@PostMapping("/addWz")
-	public ResponseData addWz(@RequestBody Wz wz) {
-		return wzService.save(wz) ? ResponseData.success().message("添加成功！")
-				: ResponseData.error().message("添加失败!");
-	}
+    @Autowired
+    private UserMapper userMapper;
 
-
-	// 删除
-	@GetMapping("/delWz/{id}")
-	public ResponseData delWz(@PathVariable("id") Integer id) {
-		return wzService.removeById(id) ? ResponseData.success().message("删除成功!")
-				: ResponseData.error().message("删除失败!");
-	}
+    // 添加
+    @PostMapping("/addWz")
+    public ResponseData addWz(@RequestBody Wz wz) {
 
 
-	// 查询 0-所有
-	@GetMapping("/getWzList/{userId}")
-	public ResponseData getWzList(@RequestParam(name = "pagenum", defaultValue = "1", required = false) long pagenum,
-									@RequestParam(name = "pagesize", defaultValue = "10", required = false) long pagesize,
-									@PathVariable("userId") Integer id) {
+        return wzService.save(wz) ? ResponseData.success().message("添加成功！")
+                : ResponseData.success().message("添加失败!");
+    }
 
-		Page<Wz> page = wzService.getInfoList(pagenum, pagesize, id);
-		if (page != null) {
-			Map<String, Object> map = new HashMap<>();
-			map.put("total", page.getTotal());
-			map.put("data", page.getRecords());
-			return ResponseData.success().message("获取成功！").data(map);
-		}
-		return ResponseData.error().message("获取失败！");
-	}
+
+    // 删除
+    @GetMapping("/delWz/{id}")
+    public ResponseData delWz(@PathVariable("id") Integer id) {
+        return wzService.removeById(id) ? ResponseData.success().message("删除成功!")
+                : ResponseData.error().message("删除失败!");
+    }
+
+
+    @GetMapping("/updWZ/{id}")
+    public ResponseData updWZ(@PathVariable("id") Integer id) {
+
+        User user = userMapper.selectById(id);
+        if (user == null) {
+            return ResponseData.error().message("没有此用户!");
+        }
+        Wz byId = wzService.getById(id);
+        userMapper.updJF(id, user.getJf() - byId.getJf());
+
+        wzMapper.upd(id);
+
+        return wzService.removeById(id) ? ResponseData.success().message("删除成功!")
+                : ResponseData.error().message("删除失败!");
+    }
+
+
+    // 查询 0-所有
+    @GetMapping("/getWzList/{userId}")
+    public ResponseData getWzList(@RequestParam(name = "pagenum", defaultValue = "1", required = false) long pagenum,
+                                  @RequestParam(name = "pagesize", defaultValue = "10", required = false) long pagesize,
+                                  @PathVariable("userId") Integer id) {
+
+        Page<WzVo> page = wzService.getInfoList(pagenum, pagesize, id);
+        if (page != null) {
+            Map<String, Object> map = new HashMap<>();
+            map.put("total", page.getTotal());
+            map.put("data", page.getRecords());
+            return ResponseData.success().message("获取成功！").data(map);
+        }
+        return ResponseData.error().message("获取失败！");
+    }
 
 }
